personal-news-digest-agent/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
feedparser
google-generativeai
